revert return type of days_in_current_year and days_in_next_yearto float and use epoch time to more accurately calaculate number of days

+++ pywr/_core.pxd
@@ -37,8 +37,8 @@ cdef class Timestep:
     cdef readonly int month
     cdef readonly int year
     cdef readonly int end_year
-    cpdef public int days_in_current_year(self)
-    cpdef public int days_in_next_year(self)
+    cpdef public double days_in_current_year(self)
+    cpdef public double days_in_next_year(self)
 
 cdef class Domain:
     cdef object name
